requirements.txt
# Smart Energy Forecasting Platform - Dependencies
# Stage 1.5: Forecast + Explain

# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Web Framework
streamlit>=1.28.0

# Visualization
plotly>=5.17.0

# HTTP Requests (for Weather API)
requests>=2.31.0

# Date/Time Utilities
holidays>=0.34

# Model Persistence
# pickle is built-in to Python, no need to install

# Optional: For advanced UTCI calculations (if needed)
# pythermalcomfort>=0.5.0

# RAG and AI Dependencies
faiss-cpu>=1.7.4
openai>=1.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0